% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbea_internals.R
\name{adjust_scores}
\alias{adjust_scores}
\title{Function to compute adjusted CBEA scores if raw scores are not the
intended output}
\usage{
adjust_scores(
  ab_tab,
  adj,
  distr,
  init,
  control,
  raw_scores,
  true_index,
  n_perm,
  parametric,
  thresh
)
}
\arguments{
\item{ab_tab}{(Matrix). Named \code{n} by \code{p} matrix.
This is the OTU/ASV/Strain table where taxa are columns.}

\item{adj}{(Logical). See documentation \code{\link{cbea}}}

\item{distr}{(Character). See documentation \code{\link{cbea}}}

\item{init}{(List). See documentation \code{\link{cbea}}}

\item{control}{(List). See documentation \code{\link{cbea}}}

\item{raw_scores}{(Vector). A vector of the raw scores using the correct
indices}

\item{true_index}{(Vector). A vector of the correct indices}

\item{n_perm}{(Numeric). The total number of permutations.}

\item{parametric}{(Logical). See documentation \code{\link{cbea}}}

\item{thresh}{(Numeric). See documentation \code{\link{cbea}}}
}
\description{
Function to compute adjusted CBEA scores if raw scores are not the
intended output
}
\keyword{internal}
