% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbea_methods.R
\name{cbea}
\alias{cbea}
\alias{cbea,phyloseq-method}
\alias{cbea,TreeSummarizedExperiment-method}
\alias{cbea,data.frame-method}
\alias{cbea,matrix-method}
\title{Enrichment analysis using competitive compositional balances (CBEA)}
\usage{
cbea(
  obj,
  set,
  output,
  distr,
  adj = FALSE,
  n_perm = 1,
  parametric = TRUE,
  thresh = 0.05,
  init = NULL,
  control = NULL,
  ...
)

\S4method{cbea}{phyloseq}(
  obj,
  set,
  output,
  distr,
  adj = FALSE,
  n_perm = 1,
  parametric = TRUE,
  thresh = 0.05,
  init = NULL,
  control = NULL,
  ...
)

\S4method{cbea}{TreeSummarizedExperiment}(
  obj,
  set,
  output,
  distr,
  adj = FALSE,
  n_perm = 1,
  parametric = TRUE,
  thresh = 0.05,
  init = NULL,
  control = NULL,
  ...
)

\S4method{cbea}{data.frame}(
  obj,
  set,
  output,
  distr,
  adj = FALSE,
  n_perm = 1,
  parametric = TRUE,
  thresh = 0.05,
  init = NULL,
  control = NULL,
  taxa_are_rows = FALSE,
  ...
)

\S4method{cbea}{matrix}(
  obj,
  set,
  output,
  distr,
  adj = FALSE,
  n_perm = 1,
  parametric = TRUE,
  thresh = 0.05,
  init = NULL,
  control = NULL,
  taxa_are_rows = FALSE,
  ...
)
}
\arguments{
\item{obj}{The element of class \code{TreeSummarizedExperiment},
\code{phyloseq}, \code{data.frame}, or \code{matrix}}

\item{set}{\code{BiocSet}. Sets to be tested for
enrichment in the \code{BiocSet}
format. Taxa names must be in the same format
as elements in the set.}

\item{output}{(String). The form of the output of the model.
Has to be either \code{zscore},
\code{cdf}, \code{raw}, \code{pval}, or \code{sig}}

\item{distr}{(String). The choice of distribution for the null. Can be either \code{mnorm}
(2 component mixture normal), \code{norm} (Normal distribution), or NULL if \code{parametric}
is \code{TRUE}.}

\item{adj}{(Logical). Whether correlation adjustment procedure is utilized. Defaults to FALSE.}

\item{n_perm}{(Numeric). Add bootstrap resamples to both the permuted and unpermuted data set.
This might help with stabilizing the distribution fitting procedure, especially if the sample
size is low. Defaults to 1.}

\item{parametric}{(Logical). Indicate whether a parametric distribution will be fitted to estimate
z-scores, CDF values, and p-values. Defaults to \code{TRUE}}

\item{thresh}{(Numeric). Threshold for significant p-values if \code{sig}
is the output. Defaults to 0.05}

\item{init}{(Named List). Initialization parameters
for estimating the null distribution.
Default is NULL.}

\item{control}{(Named List). Additional arguments to be passed to
\code{fitdistr} and \code{normmixEM}. Defaults to NULL.}

\item{...}{Additional arguments not used at the moment.}

\item{taxa_are_rows}{(Logical). Indicate whether the data frame
or matrix has taxa as rows}
}
\value{
\code{R}    An \code{n} by \code{m} matrix of enrichment scores at
the sample level
}
\description{
\code{cbea} is used compute enrichment scores
per sample for pre-defined sets using the CBEA
(Competitive Balances for Enrichment Analysis).
}
\details{
This function support different formats of the OTU table, however
for best results please use either \code{\linkS4class{TreeSummarizedExperiment}} or
\code{\linkS4class{phyloseq}}. If use \code{data.frame} or \code{matrix}, additional
arguments specifying if \code{taxa_are_rows} should be true or not. \cr
The \code{output} specifies what type of values will be returned in the final matrix.
The options \code{pval} or \code{sig} returns either unadjusted p-values or dummy variables
indicating whether a set is significantly enriched in that sample (based on unadjusted
p-values thresholded at \code{thresh}).
The option \code{raw} returns raw scores computed for each set without any distribution fitting
or inference procedure. Users can use this option to examine the distribution of CBEA scores
under the null.
}
\examples{
data(hmp_gingival)
seq <- hmp_gingival$data
set <- hmp_gingival$set
mod <- cbea(obj = seq, set = set, output = "zscore", distr = "norm", parametric = TRUE,
    adj = TRUE, thresh = 0.05)
}
