% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_args}
\alias{check_args}
\title{Checking arguments of the function}
\usage{
check_args()
}
\value{
None
}
\description{
This function extracts the parent environment (when called under the cbea function)
and then check all the arguments.
}
\keyword{internal}
