% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbea_internals.R
\name{check_distr_arg}
\alias{check_distr_arg}
\title{This function checks for validity of arguments based on
the parameters and the distribution of interest}
\usage{
check_distr_arg(param, distr, .note = NULL)
}
\arguments{
\item{param}{(List). Named list of parameter values}

\item{distr}{(String). String name of the distribution being
evaluated}

\item{.note}{(String). Any additional annotation to be put
in front of error messages}
}
\value{
Returns 0 if there are no errors
}
\description{
This function checks for validity of arguments based on
the parameters and the distribution of interest
}
\keyword{internal}
