% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cilr.R
\name{cilr}
\alias{cilr}
\title{Main cILR function to perform enrichment analysis using cILR method}
\usage{
cilr(
  X,
  A,
  output = c("cdf", "zscore", "pval", "sig"),
  distr = c("mnorm", "norm"),
  adj = TRUE,
  nperm = 5,
  thresh = 0.05,
  init = NULL,
  min_size = 2,
  raw = FALSE,
  ...
)
}
\arguments{
\item{X}{Named \code{n} by \code{p} matrix. This is the OTU/ASV/Strain table where taxa are columns.}

\item{A}{List of length \code{m}. This is a list of set membership by column names.}

\item{output}{String. The form of the output of the model.}

\item{distr}{String. The choice of distribution for the null.}

\item{adj}{Logical. Whether correlation adjustment procedure is utilized.}

\item{nperm}{Integer. The number of permutations/bootstrap re-sampling. Recommended not to exceed 5
due to memory issues.}

\item{thresh}{Numeric. Threshold for significant p-values if \code{sig} is the output.}

\item{init}{Named List. Initialization parameters for estimating the null distribution. Default is NULL.}

\item{min_size}{Integer. The lower bound of set size. Sets of size lower than min_size will be filtered out.
Default is 2 (no singleton sets).}

\item{raw}{Logical. Whether scores are returned as raw (no parameter estimation step). Default is FALSE.}

\item{...}{Named List. Additional arguments to be passed to \code{fitdistr} and \code{normmixEM}}
}
\value{
\code{R}    An \code{n} by \code{m} matrix of enrichment scores at the sample level
}
\description{
Main cILR function to perform enrichment analysis using cILR method
}
