% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbea_internals.R
\name{combine_distr}
\alias{combine_distr}
\title{Combining two distributions}
\usage{
combine_distr(perm, unperm, distr, ...)
}
\arguments{
\item{perm}{(List). A list of parameters for permuted distribution}

\item{unperm}{(List). A list of parameters for the unpermuted distribtion}

\item{distr}{(String). Distribution of choice}
}
\value{
A list of the combined distribution form based on
the initial distribution of choice
}
\description{
Pass along handling of combining distributions to avoid
clogging up the main function
}
\keyword{internal}
