% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{const_set}
\alias{const_set}
\alias{const_set,matrix-method}
\alias{const_set,taxonomyTable-method}
\alias{const_set,data.frame-method}
\title{Construct sets from objects}
\usage{
const_set(obj, ...)

\S4method{const_set}{matrix}(obj)

\S4method{const_set}{taxonomyTable}(obj, rank = NULL)

\S4method{const_set}{data.frame}(obj)
}
\arguments{
\item{obj}{Object to convert to standard sets to be used}

\item{...}{Additional arguments to be passed}
}
\description{
This function allows the construction of taxa sets of
\code{\linkS4class{BiocSet}} type from either
\code{\linkS4class{taxonomyTable-class}},
\code{matrix}, or \code{\link[tibble]{tibble}}
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: Convert membership matrix to \code{BiocSet}

\item \code{taxonomyTable}: convert taxonomyTable to \code{BiocSet}

\item \code{data.frame}: convert
}}

