% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_construction.R
\name{conv_set}
\alias{conv_set}
\title{TODO: Convert set to different formats
This is a placeholder function}
\usage{
conv_set(format = NULL)
}
\description{
TODO: Convert set to different formats
This is a placeholder function
}
