% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{create_parameters}
\alias{create_parameters}
\title{Shorthand to create parameters}
\usage{
create_parameters(params)
}
\arguments{
\item{params}{List of parameters}
}
\description{
Shorthand to create parameters
}
\keyword{internal}
