% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dlst}
\alias{dlst}
\alias{plst}
\title{Defintions for location-scale t distribution}
\usage{
dlst(x, df = 1, mu = 0, sigma = 1, log = FALSE)

plst(q, df = 1, mu = 0, sigma = 1, log = FALSE)
}
\arguments{
\item{x, q}{The data vector}

\item{df}{Degrees of freedom}

\item{mu}{The location parameter}

\item{sigma}{The scale parameter}

\item{log}{Indicate whether probabilities are return as log}
}
\description{
Internal functions for defining the t-distribution in terms
of location-scale.
}
\section{Functions}{
\itemize{
\item \code{dlst}: Probability Density Function

\item \code{plst}: Cumulative distribution function
}}

\examples{
    val <- rnorm(10)
    dlst(val, df = 1, mu = 0, sigma = 1)
    val <- rnorm(10)
    plst(q = val, df = 1, mu = 0, sigma = 1)
}
\keyword{internal}
