% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbea_internals.R
\name{.cbea}
\alias{.cbea}
\title{Internal cbea function}
\usage{
.cbea(
  ab_tab,
  set_list,
  output,
  distr,
  adj = TRUE,
  thresh = 0.05,
  init = NULL,
  raw = FALSE,
  control = NULL,
  ...
)
}
\arguments{
\item{ab_tab}{Named \code{n} by \code{p} matrix. This is the OTU/ASV/Strain table where taxa are columns.}

\item{set_list}{List of length \code{m}. This is a list of set membership by column names.}

\item{output}{See documentation \code{\link{cbea}}}

\item{distr}{See documentation \code{\link{cbea}}}

\item{adj}{See documentation \code{\link{cbea}}}

\item{thresh}{See documentation \code{\link{cbea}}}

\item{init}{See documentation \code{\link{cbea}}}

\item{raw}{See documentation \code{\link{cbea}}}

\item{control}{See documentation \code{\link{cbea}}}

\item{...}{See documentation \code{\link{cbea}}}
}
\description{
See main function \code{cbea} documentation for more details.
}
\keyword{internal}
