% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_construction.R
\name{.const_set_taxtable}
\alias{.const_set_taxtable}
\title{Internal function to construct from a tax table converted to data.frame or matrix formats}
\usage{
.const_set_taxtable(table, rank)
}
\arguments{
\item{table}{DataFrame.The taxonomic table extracted from relevant objects}

\item{rank}{String. The rank to extract sets from}
}
\description{
Internal function to construct from a tax table converted to data.frame or matrix formats
}
\keyword{internal}
