% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cilr.R
\name{estimate_distr}
\alias{estimate_distr}
\title{Estimating distribution given a vector of scores}
\usage{
estimate_distr(data, distr = c("mnorm", "norm"), init, ...)
}
\arguments{
\item{data}{\code{Matrix} otu table in matrix format where taxa are columns}

\item{distr}{\code{string} the distribution of the function}

\item{init}{\code{list} List of starting parameters}

\item{...}{Additional arguments to normix and fitdistrplus}
}
\description{
Estimating distribution given a vector of scores
}
