% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbea_internals.R
\name{fit_scores}
\alias{fit_scores}
\title{Function to compute CBEA scores for each set}
\usage{
fit_scores(
  index_vec,
  ab_tab,
  adj,
  distr,
  output,
  n_perm,
  parametric,
  thresh,
  init,
  control
)
}
\arguments{
\item{index_vec}{(Character Vector). A character vector indicating the
elements of the set of interest}

\item{ab_tab}{(Matrix). Named \code{n} by \code{p} matrix.
This is the OTU/ASV/Strain table where taxa are columns.}

\item{adj}{(Logical). See documentation \code{\link{cbea}}}

\item{distr}{(Character). See documentation \code{\link{cbea}}}

\item{output}{(Character). See documentation \code{\link{cbea}}}

\item{n_perm}{(Numeric). The total number of permutations.}

\item{parametric}{(Logical). See documentation \code{\link{cbea}}}

\item{thresh}{(Numeric). See documentation \code{\link{cbea}}}

\item{init}{(List). See documentation \code{\link{cbea}}}

\item{control}{(List). See documentation \code{\link{cbea}}}
}
\description{
Function to compute CBEA scores for each set
}
\keyword{internal}
