% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mean}
\alias{get_mean}
\title{Get the overall mean of a two component mixture distribution}
\usage{
get_mean(mu, lambda)
}
\arguments{
\item{mu}{(Vector). A two value vector of mean values.}

\item{lambda}{(Vector). A two value vector of component mixing coefficients}
}
\value{
A numeric value representing the overall mean
}
\description{
Get the overall mean of a two component mixture distribution
}
\keyword{internal}
