% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_sd}
\alias{get_sd}
\title{Get the overall standard deviation of a two component mixture distribution}
\usage{
get_sd(sigma, mu, mean, lambda)
}
\arguments{
\item{sigma}{(Vector). A two value vector of component-wise variances}

\item{mu}{(Vector). A two value vector of mean values.}

\item{mean}{(Numeric Value). The overall mean.}

\item{lambda}{(Vector). A two value vector of component mixing coefficients}
}
\value{
A numeric value representing the overall standard deviation
}
\description{
Get the overall standard deviation of a two component mixture distribution
}
\keyword{internal}
