% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_generics.R
\name{glance.CBEAout}
\alias{glance.CBEAout}
\title{Glance at \code{CBEAout} object}
\usage{
\method{glance}{CBEAout}(x, statistic, ...)
}
\arguments{
\item{x}{An object of type \code{CBEAout}}

\item{statistic}{What type of diagnostic to return. Users can choose to return
\code{fit_diagnostic} which returns goodness of fit statistics for the
different fitted distributions (e.g. log likelihoods) while \code{fit_comparison}
returns comparisons across different distributions and raw values (and data) across the
4 l-moments.}

\item{...}{Unused, kept for consistency with generics}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} summarizing diagnostic fits per set (as row)
}
\description{
This function cleans up all diagnostics of the \code{cbea} method
(from the \code{CBEAout} object) into a nice \code{\link[tibble:tibble]{tibble::tibble()}}
}
\examples{
# load the data
data(hmp_gingival)
mod <- cbea(hmp_gingival$data, hmp_gingival$set, abund_values = "16SrRNA",
    output = "sig", distr = "norm", adj = FALSE, n_perm = 5, parametric = TRUE)
glance(mod, "fit_diagnostic")
}
