% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gmean}
\alias{gmean}
\title{Geometric mean of a vector}
\usage{
gmean(vec)
}
\arguments{
\item{vec}{A vector of values with length \code{n}}
}
\value{
A numeric value of the
geometric mean of the vector \code{vec}
}
\description{
Compute geometric mean of a vector
using \code{exp(mean(log(.x)))} format
}
\examples{
ex <- abs(rnorm(10))
gmean(ex)
}
