% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gmeanRow}
\alias{gmeanRow}
\title{Geometric mean of rows of a matrix}
\usage{
gmeanRow(X)
}
\arguments{
\item{X}{A numeric matrix with \code{n} rows and \code{p} columns}
}
\value{
A numeric vector of the geometric mean of the matrix \code{X} with length \code{n}
}
\description{
This function computes the geometric mean by row of a numeric matrix
}
