% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cilr.R
\name{merge_lists}
\alias{merge_lists}
\title{This function handles the ability to merge supplied and defaults}
\usage{
merge_lists(defaults, supplied)
}
\arguments{
\item{defaults}{(List). Default options}

\item{supplied}{(List). Supplied options}
}
\value{
A merged list
}
\description{
This function handles the ability to merge supplied and defaults
}
\keyword{internal}
