% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseqSetClass.R
\docType{class}
\name{phyloseqSet-class}
\alias{phyloseqSet-class}
\title{S4 class extending phyloseq object to also store taxon sets}
\description{
This is a simple S4 object that extends \code{\link{phyloseq}} by one extra slot
with a \code{taxon_set} object which is of class \code{\link{BiocSet}} class.
}
\section{Slots}{

\describe{
\item{\code{otu_table}}{A \code{\link{otu_table}} type from \code{\link{phyloseq}}}

\item{\code{sam_data}}{A \code{\link{sam_data}} type from \code{\link{phyloseq}}}

\item{\code{tax_table}}{A \code{\link{tax_table}} type from \code{\link{phyloseq}}}

\item{\code{phy_tree}}{A \code{\link{phy_tree}} type from \code{\link{phyloseq}}}

\item{\code{refseq}}{A \code{\link{refseq}} type from \code{\link{phyloseq}}}

\item{\code{taxon_set}}{A \code{\link{BiocSet}} object}
}}

