% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pmnorm}
\alias{pmnorm}
\alias{dmnorm}
\title{The Two Component Mixture Normal Distribution}
\usage{
pmnorm(q, mu, sigma, lambda, log = FALSE, verbose = FALSE)

dmnorm(x, mu, sigma, lambda, log = FALSE, verbose = FALSE)
}
\arguments{
\item{q, x}{(Vector). Values to calculate distributional values of.}

\item{mu}{(Vector). A two value vector of mean values.}

\item{sigma}{(Vector). A two value vector of component-wise variances}

\item{lambda}{(Vector). A two value vector of component mixing coefficients}

\item{log}{(Boolean). Whether returning probabilities are in log format}

\item{verbose}{(Boolean). Whether to return component values.}
}
\value{
A numeric value representing the probability
density value of a two-component mixture distribution
}
\description{
The Two Component Mixture Normal Distribution
}
\section{Functions}{
\itemize{
\item \code{pmnorm}: Cumulative Distribution Function

\item \code{dmnorm}: Probability Density Function
}}

\examples{
    library(mixtools)
    lambda <- c(0.7,0.3)
    mu <- c(1,2)
    sigma <- c(1,1)
    v <- rnormmix(100, lambda=lambda, mu=mu, sigma=sigma)
    pmnorm(v, lambda=lambda,mu=mu,sigma=sigma)
    dmnorm(v, lambda=lambda,mu=mu,sigma=sigma)
}
