% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBEAout.R
\name{print.CBEAout}
\alias{print.CBEAout}
\title{Print dispatch for CBEAout objects}
\usage{
\method{print}{CBEAout}(x, ...)
}
\arguments{
\item{x}{The \code{CBEAout} object}

\item{...}{Undefined arguments, keeping consistency for generics}
}
\value{
Text for printing
}
\description{
Print dispatch for CBEAout objects
}
