% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulating microbiome relative abundance data
according to zero inflated negative binomial distribution}
\usage{
simulate_data(
  n_samp = 1000,
  n_tax = 300,
  spar = 0.1,
  s_rho = 0.3,
  eff_size = 3,
  b_rho = 0,
  n_inflate = 50,
  n_sets = 1,
  prop_set_inflate = 1,
  prop_inflate = 1,
  samp_prop = 0.5,
  method = "compensation",
  vary_params = TRUE,
  parameters = NULL
)
}
\arguments{
\item{n_samp}{Number of samples}

\item{n_tax}{Number of taxa}

\item{spar}{Additive sparsity}

\item{s_rho}{Correlation within the set}

\item{eff_size}{Effect size that is a multiplier to the base mu of a sample}

\item{b_rho}{Baseline inter-taxa correlation}

\item{n_inflate}{Number of taxa with inflated counts}

\item{n_sets}{Number of sets to simulate}

\item{prop_set_inflate}{Number of proportions of sets that are inflated}

\item{prop_inflate}{The number of differentially abundant taxa within a set that is inflated}

\item{samp_prop}{The proportion of samples have an inflated taxa}

\item{method}{Method of simulation. Can be "compensation", "non-compensation" or "normal"}

\item{vary_params}{Whether we stochastically draw parameter values for each feature}

\item{parameters}{Path to rds file containing estimated values}
}
\description{
Function to simulate microbiome relative abundance data. Each taxa is simulated
using a gaussian copula with a zero-inflated negative binomial marginals. Default parameters
taken from fitting zero-inflated data to stool samples from the Human Microbiome Project.
Note: n_inflate * n_sets must be equal to n_taxa.
}
