% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseqSetClass.R
\name{taxon_set<-}
\alias{taxon_set<-}
\title{Assign a new taxon set object to \code{x}}
\usage{
taxon_set(x) <- value
}
\arguments{
\item{x}{\code{\link{phyloseqSet-class}}}

\item{value}{\code{\link{BiocSet}} class of taxa sets}
}
\description{
Assign a new taxon set object to \code{x}
}
