% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseqSetClass.R
\name{taxon_set}
\alias{taxon_set}
\title{Access taxon set object from \code{phyloseqSet} object}
\usage{
taxon_set(x)
}
\arguments{
\item{x}{\link{phyloseqSet}}
}
\description{
Access taxon set object from \code{phyloseqSet} object
}
