% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cilr.R
\name{throw_error}
\alias{throw_error}
\title{Function to throw error in the distribution fitting procedure}
\usage{
throw_error(cond)
}
\arguments{
\item{cond}{Condition of the function}
}
\value{
Throw warning and return null
}
\description{
Function to throw error in the distribution fitting procedure
}
