% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_generics.R
\name{tidy.CBEAout}
\alias{tidy.CBEAout}
\title{Tidy a CBEAout object}
\usage{
\method{tidy}{CBEAout}(x, ...)
}
\arguments{
\item{x}{A \code{CBEAout} object.}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A tidy \code{\link[tibble:tibble]{tibble::tibble()}} summarizing scores per sample per set.
}
\description{
This function takes in a CBEA type object and collects all values across all
sets and samples that were evaluated.
}
\examples{
# load the data
data(hmp_gingival)
mod <- cbea(hmp_gingival$data, hmp_gingival$set, abund_values = "16SrRNA",
    output = "sig", distr = "norm", adj = FALSE, n_perm = 5, parametric = TRUE)
tidy(mod)
}
