% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_construction.R
\name{trim_set}
\alias{trim_set}
\title{Trim sets based on user supplied criteria}
\usage{
trim_set(physeq, ...)
}
\arguments{
\item{physeq}{(\code{phyloseqSet}). This object has taxon sets that need to be trimmed}

\item{...}{(Expression). Trim criteria that would be used. For example \code{size >= 5}.}
}
\value{
A \code{phyloseqSet} object with sets and elements appropriately trimmed and filtered
}
\description{
A filter operation specific to phyloseqSet type objects for sets. Using the criteria that appears
in the \code{set} slot of the \code{BiocSet} object, users can filter out both elements, sets,
and element-sets. Usage of this function is very similar to \code{\link[dplyr]{filter}} function in the \code{dplyr}
package.
}
