% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_construction.R
\name{unify_sets}
\alias{unify_sets}
\alias{unify_sets,TreeSummarizedExperiment-method}
\alias{unify_sets,phyloseq-method}
\alias{unify_sets,data.frame-method}
\alias{unify_sets,matrix-method}
\title{Unify sets with data containers}
\usage{
unify_sets(obj, set, ...)

\S4method{unify_sets}{TreeSummarizedExperiment}(obj, set)

\S4method{unify_sets}{phyloseq}(obj, set)

\S4method{unify_sets}{data.frame}(obj, set, tax_ids = NULL, taxa_are_rows = TRUE)

\S4method{unify_sets}{matrix}(obj, set, taxa_are_rows)
}
\arguments{
\item{obj}{Data container object containing taxonomic relative abundance data with taxa and sample names}

\item{set}{(BiocSet) The set being tested against}

\item{...}{Additional arguments to be passed}

\item{tax_ids}{(String). Indicate column name with all the names of taxa. Can be "rownames" if
taxa names are stored as row names. Only used if \code{taxa_are_rows = TRUE}.}

\item{taxa_are_rows}{(Logical). Indicate whether taxa are stored as rows or columns of the data.frame/matrix.}
}
\value{
An object of type \code{\linkS4class{BiocSet}}
}
\description{
This function checks for consistencies in naming between
the data container and the set itself.
}
\details{
Extracts out elements from a \code{\linkS4class{BiocSet}}
and then compares it against taxa names from the data container
(either \code{\linkS4class{phyloseq}} or \code{\linkS4class{TreeSummarizedExperiment}}).
It then removes elements from the \code{\linkS4class{BiocSet}} that are
not in the taxa_names of the data container. Throws an error if there are no elements in BiocSet that matches
taxa names of the data container.
}
\section{Methods (by class)}{
\itemize{
\item \code{TreeSummarizedExperiment}: Unify set and data container for \code{\linkS4class{TreeSummarizedExperiment}}

\item \code{phyloseq}: Unify set and data container for \code{\linkS4class{phyloseq}}

\item \code{data.frame}: Unify set and data container for \code{data.frame}

\item \code{matrix}: Unify set and data container for \code{\linkS4class{matrix}}
}}

