% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{var_setup}
\alias{var_setup}
\title{Setting up parameter arrays for vectorized call to d/pnorm functions
for multi-component mixture distributions}
\usage{
var_setup(mu, sigma, lambda, vlen)
}
\arguments{
\item{mu}{See pmnorm documentation}

\item{sigma}{See pmnorm documentation}

\item{lambda}{See pmnorm documentation}

\item{vlen}{(Integer). Length of the \code{x} or \code{p} vector to be evaluated}
}
\value{
A list containing lambda, mu, and sigma
}
\description{
Setting up parameter arrays for vectorized call to d/pnorm functions
for multi-component mixture distributions
}
\keyword{internal}
