% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbea_methods.R
\name{cbea}
\alias{cbea}
\alias{cbea,TreeSummarizedExperiment-method}
\alias{cbea,data.frame-method}
\alias{cbea,matrix-method}
\title{Enrichment analysis using competitive compositional balances (CBEA)}
\usage{
cbea(
  obj,
  set,
  output,
  distr = NULL,
  adj = FALSE,
  n_perm = 100,
  parametric = TRUE,
  thresh = 0.05,
  init = NULL,
  control = NULL,
  parallel_backend = NULL,
  ...
)

\S4method{cbea}{TreeSummarizedExperiment}(
  obj,
  set,
  output,
  distr = NULL,
  abund_values,
  adj = FALSE,
  n_perm = 100,
  parametric = TRUE,
  thresh = 0.05,
  init = NULL,
  control = NULL,
  parallel_backend = NULL,
  ...
)

\S4method{cbea}{data.frame}(
  obj,
  set,
  taxa_are_rows = FALSE,
  id_col = NULL,
  output,
  distr = NULL,
  adj = FALSE,
  n_perm = 100,
  parametric = TRUE,
  thresh = 0.05,
  init = NULL,
  control = NULL,
  parallel_backend = NULL,
  ...
)

\S4method{cbea}{matrix}(
  obj,
  set,
  taxa_are_rows = FALSE,
  output,
  distr = NULL,
  adj = FALSE,
  n_perm = 100,
  parametric = TRUE,
  thresh = 0.05,
  init = NULL,
  control = NULL,
  parallel_backend = NULL,
  ...
)
}
\arguments{
\item{obj}{The element of class \code{TreeSummarizedExperiment},
\code{data.frame}, or \code{matrix}. \code{phyloseq} is not supported
due to conflicting dependencies and \code{TreeSummarizedExperiment} is much
more compact.}

\item{set}{\code{BiocSet}. Sets to be tested for
enrichment in the \code{BiocSet}
format. Taxa names must be in the same format
as elements in the set.}

\item{output}{(String). The form of the output of the model.
Has to be either \code{zscore},
\code{cdf}, \code{raw}, \code{pval}, or \code{sig}}

\item{distr}{(String). The choice of distribution for the null. Can be either \code{mnorm}
(2 component mixture normal), \code{norm} (Normal distribution), or NULL if \code{parametric}
is \code{TRUE}.}

\item{adj}{(Logical). Whether correlation adjustment procedure is utilized. Defaults to FALSE.}

\item{n_perm}{(Numeric). Add bootstrap resamples to both the permuted and unpermuted data set.
This might help with stabilizing the distribution fitting procedure, especially if the sample
size is low. Defaults to 1.}

\item{parametric}{(Logical). Indicate whether a parametric distribution will be fitted to estimate
z-scores, CDF values, and p-values. Defaults to \code{TRUE}}

\item{thresh}{(Numeric). Threshold for significant p-values if \code{sig}
is the output. Defaults to 0.05}

\item{init}{(Named List). Initialization parameters
for estimating the null distribution.
Default is NULL.}

\item{control}{(Named List). Additional arguments to be passed to
\code{fitdistr} and \code{normmixEM}. Defaults to NULL.}

\item{parallel_backend}{See documentation \code{\link{cbea}}}

\item{...}{Additional arguments not used at the moment.}

\item{abund_values}{(Character). Character value for selecting the \code{assay} to be
the input to \code{cbea}}

\item{taxa_are_rows}{(Logical). Indicate whether the data frame
or matrix has taxa as rows}

\item{id_col}{(Character Vector). Vector of character to indicate metadata
columns to keep (for example, \code{sample_id})}
}
\value{
A \code{CBEAout} object which is a list of lists where the outer list elements contains
numerical results and diagnostic tools. Within each element is a list of size \code{m} where \code{m} is the
number of sets evaluated. Specifically, a \code{CBEAout} object has
\itemize{
\item{\code{R}: }{A list of CBEA generated values for each set. Use \code{tidy} function to
generate a data.frame}
\item{\code{parameters}: }{The parameters of the fitted distributions for each set.}
\item{\code{diagnostic}: }{Contains log-likelihood, Anderson-Darling goodness of fit statistic,
for each distribution fit performed. The distribution types will depend on function calls. Users can
use this to compare fit against different distribution choices.}
\item{\code{fit_comparison}: }{Comparing generated values from best-fit distributions using l-moments.
For example, we would compare the l-moments of the distribution fitted against permuted data and the permuted
data itself. This is to check whether the distribution fitting process did not capture the presumed null.}
\item{\code{call_param}: }{Function call values.}
}
}
\description{
\code{cbea} is used compute enrichment scores
per sample for pre-defined sets using the CBEA
(Competitive Balances for Enrichment Analysis).
}
\details{
This function support different formats of the OTU table, however
for best results please use \code{\linkS4class{TreeSummarizedExperiment}}.
\code{phyloseq} is supported, however \code{CBEA} will not explicitly import
\code{phyloseq} package and will require users to install them separately.
If use \code{data.frame} or \code{matrix}, users should specify whether taxa are rows using the
\code{taxa_are_rows} option. Additionally, for \code{data.frame}, users can specify
metadata columns to be kept via the \code{id_col} argument. \cr
The \code{output} argument specifies what type of values will be returned in the final matrix.
The options \code{pval} or \code{sig} returns either unadjusted p-values or dummy variables
indicating whether a set is significantly enriched in that sample (based on unadjusted
p-values thresholded at \code{thresh}).
The option \code{raw} returns raw scores computed for each set without any distribution fitting
or inference procedure. Users can use this option to examine the distribution of CBEA scores
under the null.
}
\examples{
data(hmp_gingival)
seq <- hmp_gingival$data
set <- hmp_gingival$set
# n_perm = 10 to reduce runtime
mod <- cbea(obj = seq, set = set, output = "zscore",
    abund_values = "16SrRNA",
    distr = "norm", parametric = TRUE,
    adj = TRUE, thresh = 0.05, n_perm = 10)
}
