% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cilr.R
\name{cilr}
\alias{cilr}
\title{Applying the cILR method to a data.frame object and pre-defined sets}
\usage{
cilr(
  data,
  sets,
  output = NULL,
  distr = c("mnorm", "norm"),
  min_size = 2,
  nperm = 5,
  adj = TRUE,
  thresh = NULL,
  init = NULL,
  ...
)
}
\arguments{
\item{data}{\code{matrix}. The otu or asv table in data.frame format}

\item{sets}{\code{list}. The sets exists as named list with set names as strings and column index as integer}

\item{output}{\code{string}. What is the preferred output, can be either
\code{zscores}, \code{cdf}, \code{pvalues}, or \code{sig}}

\item{distr}{\code{string}. What is the distribution, required only when \code{estim} is TRUE,
can only be \code{norm} or \code{mnorm}}

\item{min_size}{\code{numeric}. What is the minimum size of a set}

\item{adj}{\code{logical}. Whether correlation adjustment should be done.}

\item{thresh}{\code{numeric}. Threshold to consider rejecting the null hypothesis based on p-values.
Only relevant when output is \code{sig}. Defaults to 0.05}

\item{init}{\code{list}. List of initialization parameters for the fitting portion, if not specified, defaults to
default options in \code{mixtools} and \code{fitdistrplus} packages}

\item{...}{Additional parameters passed to fitdistrplus or mixtools functions}

\item{n_perm}{\code{numeric}. Number of resamples for the data set for good estimation, defaults to 5.}

\item{p_adj}{\code{logical}. Whether to adjust p-values using the BH procedure. Defaults to FALSE}
}
\value{
A \code{data.frame} object with rows equals the number of samples and
    columns specifies the number of samples defined in the set
}
\description{
This function performs the cILR aggregation algorithm as defined in the Nguyen et al. paper. This
   function takes in two inputs: an otu table as a \code{matrix} object where rows are samples and columns are
   taxa; a named \code{list} where each list element is a numeric vector of indices for the columns of interest
}
