% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_construction.R
\name{const_set}
\alias{const_set}
\alias{const_set,taxonomyTable-method}
\alias{const_set,matrix-method}
\alias{const_set,data.frame-method}
\title{Construct sets from objects}
\usage{
const_set(obj, min_size, ...)

\S4method{const_set}{taxonomyTable}(obj, min_size, rank)

\S4method{const_set}{matrix}(obj)

\S4method{const_set}{data.frame}(obj)
}
\arguments{
\item{obj}{Object to convert to standard sets to be used}

\item{min_size}{The minimum size of a set}

\item{...}{Additional arguments to be passed}

\item{rank}{(Character). Restrict sets to certain taxonomic ranks.}
}
\description{
This function allows the construction of taxa sets of
\code{\linkS4class{BiocSet}} type from either
\code{\linkS4class{taxonomyTable}},
\code{matrix}, or \code{\link[tibble]{tibble}}
}
\section{Methods (by class)}{
\itemize{
\item \code{taxonomyTable}: convert taxonomyTable to \code{BiocSet}

\item \code{matrix}: Convert membership (or logical) matrix to \code{BiocSet}

\item \code{data.frame}: Convert \code{data.frame} membership or \code{tibble} to
\code{BiocSet}
}}

