% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbea.R
\name{.cbea}
\alias{.cbea}
\title{Enrichment analysis using competitive compositional balances (CBEA)}
\usage{
.cbea(
  ab_tab,
  set_list,
  output = c("cdf", "zscore", "pval", "sig"),
  distr = c("mnorm", "norm"),
  adj = TRUE,
  thresh = 0.05,
  init = NULL,
  raw = FALSE,
  ...
)
}
\arguments{
\item{ab_tab}{Named \code{n} by \code{p} matrix. This is the OTU/ASV/Strain table where taxa are columns.}

\item{set_list}{List of length \code{m}. This is a list of set membership by column names.}

\item{output}{String. The form of the output of the model.}

\item{distr}{String. The choice of distribution for the null.}

\item{adj}{Logical. Whether correlation adjustment procedure is utilized.}

\item{thresh}{Numeric. Threshold for significant p-values if \code{sig} is the output.}

\item{init}{Named List. Initialization parameters for estimating the null distribution. Default is NULL.}

\item{raw}{Logical. Whether scores are returned as raw (no parameter estimation step). Default is FALSE.}

\item{...}{Named List. Additional arguments to be passed to \code{fitdistr} and \code{normmixEM}}
}
\value{
\code{R}    An \code{n} by \code{m} matrix of enrichment scores at the sample level
}
\description{
This function is an independent implementation of CBEA approach (Nguyen et al. 2021+). Users can
use this approach using this function
}
