% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbea_internals.R
\name{.cbea}
\alias{.cbea}
\title{Internal cbea function}
\usage{
.cbea(
  ab_tab,
  set_list,
  output,
  distr,
  adj = TRUE,
  n_perm = 100,
  parametric = TRUE,
  thresh = 0.05,
  init = NULL,
  control = NULL,
  parallel_backend = NULL,
  ...
)
}
\arguments{
\item{ab_tab}{(Matrix). Named \code{n} by \code{p} matrix.
This is the OTU/ASV/Strain table where taxa are columns.}

\item{set_list}{(List). List of length \code{m}.
This is a list of set membership by column names.}

\item{output}{See documentation \code{\link{cbea}}}

\item{distr}{See documentation \code{\link{cbea}}}

\item{adj}{See documentation \code{\link{cbea}}}

\item{n_perm}{See documentation \code{\link{cbea}}}

\item{parametric}{See documentation \code{\link{cbea}}}

\item{thresh}{See documentation \code{\link{cbea}}}

\item{init}{See documentation \code{\link{cbea}}}

\item{control}{See documentation \code{\link{cbea}}}

\item{parallel_backend}{See documentation \code{\link{cbea}}}

\item{...}{See documentation \code{\link{cbea}}}
}
\value{
A \code{data.frame} of size \code{n} by \code{m}.
\code{n} is the total number of samples and \code{m}
is the total number of sets with elements represented in the data.
}
\description{
See main function \code{cbea} documentation for more details.
}
\keyword{internal}
