% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cilr.R
\name{estimate_distr}
\alias{estimate_distr}
\title{Estimate distribution parameters from data}
\usage{
estimate_distr(data, distr = c("mnorm", "norm"), init = NULL, ...)
}
\arguments{
\item{data}{(Numeric Vector). A vector of numbers that can be inputted to estimate the
parameters of the distributional forms.}

\item{distr}{(String). The distribution to be fitted. Right now only \code{norm} or \code{mnorm} is supported}

\item{init}{(List). Initialization parameters for each distribution. For mixtures, each named
element in the list should be a vector with length equal to the number of components}

\item{...}{Other paremteres passed to fitdistrplus or normalmixEM}
}
\description{
This function takes a numeric vector input and attempts to find the most optimal solution
for the parameters of the distribution of choice. Right now only \code{norm} and \code{mnorm} distributions
are supported.
}
\details{
The package \code{\link[fitdistrplus]{fitdistrplus}} is used to estimate parameters of the normal distribution
while the package \code{\link[mixtools]{normalmixEM}} is used to estimate parameters of the mixture normal
distribution. So far we suggest only estimating two components for the mixture normal distribution.
For default options, we use mostly defaults from the packages themselves. The only difference was the mixture normal distribution
where the convergence parameters were loosened and requiring more iterations to converge.
}
\keyword{internal}
