% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbea_internals.R
\name{get_adj_mnorm}
\alias{get_adj_mnorm}
\title{Function to perform the adjustment for the mixture normal distribution}
\usage{
get_adj_mnorm(perm, unperm, verbose = FALSE)
}
\arguments{
\item{perm}{(List). Parameter values of the distribution of scores}

\item{unperm}{(List). Parameter values of the distribution of scores
computed on unpermuted data}
}
\value{
A List of parameters for the adjusted mixture normal.
}
\description{
Function to perform the adjustment for the mixture normal distribution
}
\keyword{internal}
