% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cilr.R
\name{get_score}
\alias{get_score}
\title{Internal function calculate scores based on a certain index}
\usage{
get_score(data, set_idx)
}
\arguments{
\item{data}{\code{matrix}. The otu table where taxa are columns and samples are rows}

\item{set_idx}{\code{vector}. Numeric vector of index of columns belonging to a certain set}
}
\description{
Internal function calculate scores based on a certain index
}
