% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbea.R
\name{get_score}
\alias{get_score}
\title{Get CBEA scores for a given matrix and a vector of column indices}
\usage{
get_score(X, idx)
}
\arguments{
\item{X}{(Matrix). OTU table of matrix format where taxa are columns and samples are rows}

\item{idx}{(Integer vector). Vector of integers indicating the column ids of taxa in a set}
}
\description{
Get CBEA scores for a given matrix and a vector of column indices
}
