% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbea_output.R
\name{new_CBEAout}
\alias{new_CBEAout}
\title{Creating an output object of type CBEAout}
\usage{
new_CBEAout(out, sample_ids, output)
}
\arguments{
\item{out}{A list containing scores for each set}
}
\description{
This function takes a list of lists from each object and
turns it into a CBEAout type object
}
