% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBEAout.R
\name{new_CBEAout}
\alias{new_CBEAout}
\title{Creating an output object of type CBEAout}
\usage{
new_CBEAout(out, call)
}
\arguments{
\item{out}{A list containing scores for each set}

\item{call}{A list containing all important arguments for printing}
}
\value{
A new CBEAout object (which is a cleaner list of lists)
}
\description{
This function takes a list of lists from each object and
turns it into a CBEAout type object
}
